requirements.txt
beautifulsoup4==4.13.4
pandas==2.3.1
PyYAML==6.0.2